+++ bin/faucet/src/faucet/mod.rs
@@ -1,8 +1,11 @@
-use std::{collections::VecDeque, rc::Rc, sync::Arc};
+use std::{collections::VecDeque, sync::Arc};
 
 use anyhow::{Context, anyhow};
 use miden_lib::{
-    account::interface::{AccountInterface, AccountInterfaceError},
+    account::{
+        faucets::BasicFungibleFaucet,
+        interface::{AccountInterface, AccountInterfaceError},
+    },
     note::create_p2id_note,
 };
 use miden_objects::{
@@ -159,13 +162,14 @@ pub struct Faucet {
     // Previous faucet account states used to perform rollbacks if a desync is detected.
     prior_state: VecDeque<Account>,
     tx_prover: Arc<FaucetProver>,
-    tx_executor: Rc<TransactionExecutor>,
+    authenticator: BasicAuthenticator<StdRng>,
     account_interface: AccountInterface,
+    decimals: u8,
 }
 
 impl Faucet {
     /// Loads the faucet state from the node and the account file.
-    #[instrument(name = "Faucet::load", fields(id), skip_all)]
+    #[instrument(name = "faucet.load", fields(id), skip_all)]
     pub async fn load(
         account_file: AccountFile,
         rpc_client: &mut RpcClient,
@@ -217,6 +221,7 @@ impl Faucet {
         )
         .expect("Empty ChainMmr should be valid");
 
+        let faucet = BasicFungibleFaucet::try_from(&account)?;
         let account_interface = AccountInterface::from(&account);
 
         let data_store = Arc::new(FaucetDataStore::new(
@@ -226,30 +231,31 @@ impl Faucet {
             genesis_chain_mmr,
         ));
 
-        let public_key = match &account_file.auth_secret_key {
-            AuthSecretKey::RpoFalcon512(secret) => secret.public_key(),
+        // We expect exactly one secret key. Anything else implies a bug elsewhere.
+        let [keypair] = account_file.auth_secret_keys.as_slice() else {
+            anyhow::bail!(
+                "Account file must contain exactly 1 authentication key, but found {}",
+                account_file.auth_secret_keys.len()
+            );
         };
-
-        let authenticator = Arc::new(BasicAuthenticator::<StdRng>::new(&[(
-            public_key.into(),
-            account_file.auth_secret_key,
-        )]));
+        let keypair = match keypair {
+            AuthSecretKey::RpoFalcon512(falcon) => (falcon.public_key().into(), keypair.clone()),
+        };
+        let authenticator = BasicAuthenticator::<StdRng>::new(&[keypair]);
 
         let tx_prover = match remote_tx_prover_url {
             Some(url) => Arc::new(FaucetProver::remote(url)),
             None => Arc::new(FaucetProver::local()),
         };
 
-        let tx_executor =
-            Rc::new(TransactionExecutor::new(data_store.clone(), Some(authenticator.clone())));
-
         Ok(Self {
             data_store,
             id,
             prior_state: VecDeque::new(),
             tx_prover,
-            tx_executor,
             account_interface,
+            decimals: faucet.decimals(),
+            authenticator,
         })
     }
 
@@ -384,7 +390,7 @@ impl Faucet {
 
         let mut rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
 
-        let p2id_notes = P2IdNotes::build(self.faucet_id(), requests, &mut rng)?;
+        let p2id_notes = P2IdNotes::build(self.faucet_id(), self.decimals, requests, &mut rng)?;
 
         // Build the note
         let notes = p2id_notes.into_inner();
@@ -431,7 +437,9 @@ impl Faucet {
         &self,
         tx_args: TransactionArgs,
     ) -> MintResult<ExecutedTransaction> {
-        self.tx_executor
+        let executor =
+            TransactionExecutor::new(self.data_store.as_ref(), Some(&self.authenticator));
+        executor
             .execute_transaction(
                 self.id.inner(),
                 BlockNumber::GENESIS,
@@ -488,6 +496,7 @@ impl P2IdNotes {
     /// Returns an error if creating any p2id note fails.
     fn build(
         source: FaucetId,
+        decimals: u8,
         requests: &[MintRequest],
         rng: &mut RpoRandomCoin,
     ) -> Result<Self, MintError> {
@@ -495,8 +504,9 @@ impl P2IdNotes {
         // ids are validated on the request level.
         let mut notes = Vec::new();
         for request in requests {
+            let amount = request.asset_amount.inner() * 10u64.pow(decimals.into());
             // SAFETY: source is definitely a faucet account, and the amount is valid.
-            let asset = FungibleAsset::new(source.inner(), request.asset_amount.inner()).unwrap();
+            let asset = FungibleAsset::new(source.inner(), amount).unwrap();
             let note = create_p2id_note(
                 source.inner(),
                 request.account_id,
@@ -517,7 +527,7 @@ impl P2IdNotes {
 
 #[cfg(test)]
 mod tests {
-    use std::{str::FromStr, sync::Mutex};
+    use std::{str::FromStr, sync::Mutex, time::Duration};
 
     use miden_lib::{AuthScheme, account::faucets::create_basic_fungible_faucet};
     use miden_node_block_producer::errors::{AddTransactionError, VerifyTxError};
@@ -529,6 +539,7 @@ mod tests {
     };
     use rand::{Rng, SeedableRng};
     use rand_chacha::ChaCha20Rng;
+    use tokio::time::{Instant, sleep};
     use url::Url;
 
     use super::*;
@@ -565,23 +576,31 @@ mod tests {
         // Start the stub node
         tokio::spawn(async move { serve_stub(&stub_node_url).await.unwrap() });
 
+        // Wait for the stub node to serve requests
+        let start = Instant::now();
+        while rpc_client.get_genesis_header().await.is_err() {
+            sleep(Duration::from_millis(100)).await;
+            assert!(start.elapsed() < Duration::from_secs(5), "stub node took too long to start");
+        }
+
         // Create the faucet
         let faucet = {
-            let genesis_header = rpc_client.get_genesis_header().await.unwrap();
             let mut rng = ChaCha20Rng::from_seed(rand::random());
             let secret = SecretKey::with_rng(&mut get_rpo_random_coin(&mut rng));
             let (account, account_seed) = create_basic_fungible_faucet(
                 rng.random(),
-                (&genesis_header).try_into().unwrap(),
                 TokenSymbol::try_from("POL").unwrap(),
                 2,
-                Felt::from(1_000_000_u32),
+                Felt::try_from(1_000_000_000_000u64).unwrap(),
                 AccountStorageMode::Public,
                 AuthScheme::RpoFalcon512 { pub_key: secret.public_key() },
             )
             .unwrap();
-            let account_file =
-                AccountFile::new(account, Some(account_seed), AuthSecretKey::RpoFalcon512(secret));
+            let account_file = AccountFile::new(
+                account,
+                Some(account_seed),
+                vec![AuthSecretKey::RpoFalcon512(secret)],
+            );
 
             Faucet::load(account_file, &mut rpc_client, None).await.unwrap()
         };
@@ -609,7 +628,9 @@ mod tests {
         let mut rng = *rng.lock().unwrap();
 
         // Build and execute the transaction
-        let notes = P2IdNotes::build(faucet.faucet_id(), &requests, &mut rng).unwrap().into_inner();
+        let notes = P2IdNotes::build(faucet.faucet_id(), 6, &requests, &mut rng)
+            .unwrap()
+            .into_inner();
         let tx_args = faucet.compile(&notes).unwrap();
 
         faucet